Time: 0 ms (100%), Space: 13.1 MB (78.35%) - LeetHub

+++ 2692-take-gifts-from-the-richest-pile/2692-take-gifts-from-the-richest-pile.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long pickGifts(vector<int>& gifts, int k) {
-        priority_queue<int> pq;
+        priority_queue<long long> pq;
         for(auto it:gifts) pq.push(it);
         while(k>0){
             k--;
@@ -10,7 +10,7 @@ public:
             n=sqrt(n);
             pq.push(n);
         }
-        int ans=0;
+        long long ans=0;
         while(!pq.empty()){
             ans+=pq.top();
             pq.pop();
